% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-brier.R
\name{brier_surv}
\alias{brier_surv}
\alias{brier_surv.data.frame}
\alias{brier_surv_vec}
\title{Brier score for censored data models}
\usage{
brier_surv(data, ...)

\method{brier_surv}{data.frame}(
  data,
  truth,
  estimate,
  .time,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

brier_surv_vec(truth, estimate, .time, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected.
Otherwise, there should be as many columns as factor levels of \code{truth}.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{If \code{truth} is binary, a numeric vector of class probabilities
corresponding to the "relevant" class. Otherwise, a matrix with as many
columns as factor levels of \code{truth}. \emph{It is assumed that these are in the
same order as the levels of \code{truth}.}}

\item{.time}{A vector of time points.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{brier_surv_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Compute the Brier score for a censored data model.
}
\details{
Smaller values of the score are associated with better model performance.
}
\section{Multiclass}{

Brier scores can be computed in the same way for any number of classes.
Because of this, no averaging types are supported.
}

\author{
Emil Hvitfeldt
}
\concept{dynamic survival metrics}
