% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-type-docs.R
\name{class-metrics}
\alias{class-metrics}
\title{Class metrics}
\description{
Class metrics evaluate hard classification predictions where both \code{truth}
and \code{estimate} are factors. These metrics compare predicted classes directly
against the true classes.
}
\section{Input requirements}{

\itemize{
\item \code{truth}: factor
\item \code{estimate}: factor
}
}

\section{Available metrics}{

\describe{
\item{\code{\link[=accuracy]{accuracy()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=bal_accuracy]{bal_accuracy()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=detection_prevalence]{detection_prevalence()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=f_meas]{f_meas()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=fall_out]{fall_out()}}}{Direction: minimize. Range: [0, 1]}
\item{\code{\link[=j_index]{j_index()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=kap]{kap()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=markedness]{markedness()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=mcc]{mcc()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=miss_rate]{miss_rate()}}}{Direction: minimize. Range: [0, 1]}
\item{\code{\link[=npv]{npv()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=ppv]{ppv()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=precision]{precision()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=recall]{recall()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=roc_dist]{roc_dist()}}}{Direction: minimize. Range: [0, 1.4142135623731]}
\item{\code{\link[=sens]{sens()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=sensitivity]{sensitivity()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=spec]{spec()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=specificity]{specificity()}}}{Direction: maximize. Range: [0, 1]}
}
}

\examples{
data("two_class_example")

head(two_class_example)

accuracy(two_class_example, truth, predicted)

}
\seealso{
\link{prob-metrics} for class probability metrics

\link{ordered-prob-metrics} for ordered probability metrics

\code{vignette("metric-types")} for an overview of all metric types
}
\keyword{internal}
