% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordance}
\alias{concordance}
\alias{concordance.data.frame}
\title{Calculate concordance, censoring-adjusted concordance and censoring-adjusted AUC for survival models}
\usage{
concordance(data, ...)

\method{concordance}{data.frame}(data, truth, estimate, training = NULL,
  invert = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{data}{For the default functions, a factor containing the
discrete measurements. For the \code{table} or \code{matrix}
functions, a table or matrix object, respectively, where the
true class results should be in the columns of the table.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a Surv object). This should an unquoted column name although
this argument is passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{estimate}{The column identifier for the predicted hazard or predicted
survival time (that is numeric). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name.}

\item{training}{Surv object corresponding to the data used to train the
model use to generate the predictions. Used to adjust the concordance
statistic for the amount of censoring in the training data.}

\item{invert}{A logical value, which controls whether or not the ordering of
the "estimate" column is inverted before concordance or AUC is calculated.
Importantly, this should be set to TRUE for predictions of hazard,
and FALSE for predictions of survival time.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be stripped before the computation proceeds}
}
\value{
A number between 0 and 1 (or NA).
}
\description{
These functions calculate the
}
\details{
The concordance is defined as the

The censoring-adjusted concordance is defined as the

The survival AUC predictive value is defined as the

See the references for discussions of the statistics.
}
\examples{
data("surv_example")

# Given that a sample is Class 1,
#   what is the probability that is predicted as Class 1?
sens(two_class_example, truth = truth, estimate = predicted)

# Given that a sample is predicted to be Class 1,
#  what is the probability that it truly is Class 1?
ppv(two_class_example, truth = truth, estimate = predicted)

# But what if we think that Class 1 only occurs 40\% of the time?
ppv(two_class_example, truth, predicted, prevalence = 0.40)
}
\references{
Harrell, F. E., R. M. Califf, D. B. Pryor, K. L. Lee and R. A.
Rosati (1982).
Evaluating the yield of medical tests.
\emph{Journal of the American Medical Association} \emph{247}, 2543-2546.

Harrell, F. E., K. L. Lee, R. M. Califf, D. B. Pryor and R. A.
Rosati (1984).
Regression modeling strategies for improved prognostic prediction.
\emph{Statistics in Medicine} \emph{3}, 143-152.

Uno, H., T. Cai T, M. J. Pencina, R. B. D'Agostino and W. L. Wei
(2011).
On the C-statistics for evaluating overall adequacy of risk
prediction procedures with censored survival data.
\emph{Statistics in Medicine} \emph{30}, 1105-1117.

Uno H, Cai T, Pencina MJ, D’Agostino RB, Wei LJ.
On the C-statistics for Evaluating Overall Adequacy of Risk Prediction
Procedures with Censored Survival Data. Statistics in medicine.
2011;30(10):1105-1117. doi:10.1002/sim.4154.

Uno, H., T. Cai, L. Tian, and L. J. Wei (2007).
Evaluating prediction rules for t-year survivors with censored
regression models.
\emph{Journal of the American Statistical Association} \emph{102}, 527-537.
}
\seealso{
survival::survConcordance survAUC::UnoC survAUC::AUC.uno
}
\keyword{manip}
