% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging-helpers.R, R/probs.R
\name{finalize_averaging}
\alias{finalize_averaging}
\alias{developer-helpers}
\alias{dots_to_estimate}
\title{Developer helpers}
\usage{
finalize_averaging(x, averaging)

dots_to_estimate(data, ...)
}
\arguments{
\item{x}{An object to dispatch the auto-selection off of. This is generally
the \code{truth} column.}

\item{averaging}{Either \code{NULL} for auto-selection, or a single character
for the type of averaging to use.}

\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate}
columns.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected.
Otherwise, there should be as many columns as factor levels of \code{truth}.}
}
\description{
Helpers to be used alongside \code{\link[=metric_vec_template]{metric_vec_template()}} and \code{\link[=metric_summarizer]{metric_summarizer()}}
when creating new metrics.
}
\details{
\code{dots_to_estimate()} is useful with class probability metrics that take
\code{...} rather than \code{estimate} as an argument. It constructs either a single
name if 1 input is provided to \code{...} or it constructs a quosure where the
expression constructs a matrix of as many columns as are provided to \code{...}.
These are eventually evaluated in the \code{summarise()} call in
\code{\link[=metric_summarizer]{metric_summarizer()}} and evaluate to either a vector or a matrix for further
use.

\code{finalize_averaging()} is the engine for auto-selection of averaging based
on the type of \code{x}. Generally \code{x} is the \code{truth} column.
If \code{averaging} is not \code{NULL}, it is returned immediately
as no auto-selection is needed. If \code{x} is a:
\itemize{
\item \code{factor} - Then \code{"binary"} is returned if it has 2 levels, otherwise
\code{"macro"} is returned.
\item \code{numeric} - Then \code{"binary"} is returned.
\item \code{table} - Then \code{"binary"} is returned if it has 2 columns, otherwise
\code{"macro"} is returned. This is useful if you have \code{table} methods.
\item \code{matrix} - Then \code{"macro"} is returned.
}
}
\seealso{
\code{\link[=metric_summarizer]{metric_summarizer()}} \code{\link[=metric_vec_template]{metric_vec_template()}}
}
