% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair-aaa.R
\name{fairness_metric}
\alias{fairness_metric}
\title{Create fairness metrics}
\usage{
fairness_metric(.fn, .name, .post)
}
\arguments{
\item{.fn}{A yardstick metric function or metric set.}

\item{.name}{The name of the metric to place in the \code{.metric} column
of the output.}

\item{.post}{A function to post-process the generated metric set results \code{x}.
In many cases, \code{~diff(range(x$.estimate))} or
\verb{~r <- range(x$.estimate); r[1]/r[2]}.}
}
\value{
A function with one argument, \code{by}, indicating the data-masked
variable giving the sensitive feature. See the documentation on the
return value of fairness metrics like \code{\link[=demographic_parity]{demographic_parity()}},
\code{\link[=equal_opportunity]{equal_opportunity()}}, or \code{\link[=equalized_odds]{equalized_odds()}} to learn more about how the
output of this function can be used.
}
\description{
Fairness metrics quantify the disparity in value of a metric across a number
of groups. Fairness metrics with a value of zero indicate that the
underlying metric has parity across groups. yardstick defines
several common fairness metrics using this function, such as
\code{\link[=demographic_parity]{demographic_parity()}}, \code{\link[=equal_opportunity]{equal_opportunity()}}, and \code{\link[=equalized_odds]{equalized_odds()}}.
}
\section{Relevant Group Level}{

Additional arguments can be passed to the function outputted by
the function that this function outputs. That is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{res_fairness <- fairness_metric(...)
res_by <- res_fairness(by)
res_by(..., additional_arguments_to_.post = TRUE)
}\if{html}{\out{</div>}}

For finer control of how groups in \code{by} are treated, use the
\code{.post} argument.
}

\examples{
data(hpc_cv)

# `demographic_parity`, among other fairness metrics,
# is generated with `fairness_metric()`:
diff_range <- function(x, ...) {diff(range(x$.estimate))}
demographic_parity_ <-
  fairness_metric(
    .fn = detection_prevalence,
    .name = "demographic_parity",
    .post = diff_range
  )

m_set <- metric_set(demographic_parity_(Resample))

m_set(hpc_cv, truth = obs, estimate = pred)

# the `post` argument can be used to accommodate a wide
# variety of parameterizations. to encode demographic
# parity as a ratio inside of a difference, for example:
ratio_range <- function(x, ...) {
  range <- range(x$.estimate)
  range[1] / range[2]
}

demographic_parity_ratio <-
  fairness_metric(
    .fn = detection_prevalence,
    .name = "demographic_parity_ratio",
    .post = ratio_range
  )

}
