% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{get_cindex}
\alias{get_cindex}
\alias{get_cindex}
\title{Get c-index}
\usage{
get_cindex(data, mod, ...)

get_cindex(data, mod, ...)
}
\arguments{
\item{data}{For the default functions, a datframe containing survival
(time), and status (0:censored/1:event), and the explanatory variables.}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
A cindex object
}
\description{
The efficacy of the survival model can be measured by
the concordance statistic
}
\examples{
data("surv_example")

# Given a sample
require(survival)
mod <- coxph(Surv(time, status)~ age, data = surv_example)

cindex(surv_example, mod)

}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
\emph{Modeling Survival Data: Extending the Cox Model}.
Springer, New York. ISBN 0-387-98784-3.
@export get_cindex
}
\seealso{
\link{coxph}
}
\author{
Carlos S Traynor
}
\keyword{cindex}
