% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_deviance.R
\name{get_deviance}
\alias{get_deviance}
\title{Get Deviance}
\usage{
get_deviance(mod, ...)
}
\arguments{
\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
Deviance
}
\description{
Information theory stablishes deviance as a measure of
uncertainty reduction. This function can be used to estimate the
out of sample deviance. Deviance should not be used when the
distribution is not the same across compared models.
}
\examples{
data("surv_example")

# Given a sample
require(survival)
mod <- coxph(Surv(time, status)~ age, data = surv_example)

get_deviance(surv_example, mod)
}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
\emph{Modeling Survival Data: Extending the Cox Model}.
Springer, New York. ISBN 0-387-98784-3.
}
\seealso{
\link{coxph}
}
\author{
Carlos S Traynor
}
\keyword{deviance}
