% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get all metrics of a given type}
\usage{
get_metrics(type)
}
\arguments{
\item{type}{A character vector of metric types. Valid types are: \code{"class"},
\code{"prob"}, \code{"ordered_prob"}, \code{"numeric"}, \code{"dynamic_survival"},
\code{"integrated_survival"}, \code{"static_survival"}, \code{"linear_pred_survival"},
and \code{"quantile"}.

More than 1 type can be selected but you are constrained by which metric
types \code{\link[=metric_set]{metric_set()}} allows to be combined. This means that
\code{get_metrics(c("class", "prob"))}  will run without error, but
\code{get_metrics(c("class", "numeric"))} will return an error because you can't
combine \code{"class"} and \code{"numeric"} metrics.}
}
\value{
A \code{\link[=metric_set]{metric_set()}} containing all metrics of the specified type(s).
}
\description{
\code{get_metrics()} returns a \code{\link[=metric_set]{metric_set()}} containing all yardstick metrics of
the specified type(s).
}
\examples{
get_metrics("numeric")

get_metrics("class")

# Get multiple types at once
get_metrics(c("class", "prob"))

}
\seealso{
\code{\link[=metric_set]{metric_set()}}
}
