% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missings.R
\name{handle_missings}
\alias{handle_missings}
\alias{detect_missings}
\title{Developer function for handling missing values in new metrics}
\usage{
handle_missings(truth, estimate, case_weights)

detect_missings(truth, estimate, case_weights)
}
\arguments{
\item{truth}{The realized vector of \code{truth}. This is either a factor or a
numeric.}

\item{estimate}{The realized \code{estimate} result. This is either a numeric
vector, a factor vector, or a numeric matrix (in the case of multiple class
probability columns) depending on your metric function.}

\item{case_weights}{The realized case weights, as a numeric vector. This must
be the same length as \code{truth}, and will be considered in the \code{na_rm}
checks. This will be passed on to \code{metric_impl} as the named argument
\code{case_weights}.}
}
\description{
\code{handle_missings()},  and \code{detect_missings()} are useful alongside the
\code{\link[=metric-summarizers]{metric-summarizers()}} functions for implementing new custom metrics.
\code{handle_missings()} removes any observations that contains missing values
across, truth, estimate and case_weights. \code{detect_missings()} returns an
indicator if there is any missing values in the inputs.
}
\seealso{
\code{\link[=metric-summarizers]{metric-summarizers()}}
}
