% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-type-docs.R
\name{linear-pred-survival-metrics}
\alias{linear-pred-survival-metrics}
\title{Linear predictor survival metrics}
\description{
Linear predictor survival metrics evaluate survival model predictions based
on linear predictors (log-hazard or log-risk scores).
}
\section{Input requirements}{

\itemize{
\item \code{truth}: a \code{\link[survival:Surv]{survival::Surv()}} object
\item \code{estimate}: numeric (linear predictor values)
}
}

\section{Available metrics}{

\describe{
\item{\code{\link[=royston_survival]{royston_survival()}}}{Direction: maximize. Range: [0, 1]}
}
}

\examples{
\dontshow{if (rlang::is_installed("tidyr")) withAutoprint(\{ # examplesIf}
data("lung_surv")

head(lung_surv)

lung_surv |>
  royston_survival(truth = surv_obj, estimate = .pred_linear_pred)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{dynamic-survival-metrics} for time-dependent survival metrics

\link{integrated-survival-metrics} for integrated survival metrics

\link{static-survival-metrics} for static survival metrics

\code{vignette("metric-types")} for an overview of all metric types
}
\keyword{internal}
