% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_surv}
\alias{lung_surv}
\title{Survival Analysis Results}
\value{
\item{lung_surv}{a data frame}
}
\description{
Survival Analysis Results
}
\details{
This data contain plausible results, when applying predictive
survival models to the \link{lung} data set using the censored package.
}
\examples{
data(lung_surv)
str(lung_surv)

# `surv_obj` is a `Surv()` object
}
\keyword{datasets}
