% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-npv.R
\name{npv}
\alias{npv}
\alias{npv.data.frame}
\alias{npv_vec}
\title{Negative predictive value}
\usage{
npv(data, ...)

\method{npv}{data.frame}(
  data,
  truth,
  estimate,
  prevalence = NULL,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)

npv_vec(
  truth,
  estimate,
  prevalence = NULL,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = yardstick_event_level(),
  ...
)
}
\arguments{
\item{data}{Either a \code{data.frame} containing the columns specified by the
\code{truth} and \code{estimate} arguments, or a \code{table}/\code{matrix} where the true
class results should be in the columns of the table.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{factor} vector.}

\item{prevalence}{A numeric value for the rate of the
"positive" class of the data.}

\item{estimator}{One of: \code{"binary"}, \code{"macro"}, \code{"macro_weighted"},
or \code{"micro"} to specify the type of averaging to be done. \code{"binary"} is
only relevant for the two class case. The other three are general methods
for calculating multiclass metrics. The default will automatically choose
\code{"binary"} or \code{"macro"} based on \code{estimate}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}

\item{event_level}{A single string. Either \code{"first"} or \code{"second"} to specify
which level of \code{truth} to consider as the "event". This argument is only
applicable when \code{estimator = "binary"}. The default uses an internal helper
that defaults to \code{"first"}.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{npv_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
These functions calculate the \code{\link[=npv]{npv()}} (negative predictive value) of a
measurement system compared to a reference result (the "truth" or gold standard).
Highly related functions are \code{\link[=spec]{spec()}}, \code{\link[=sens]{sens()}}, and \code{\link[=ppv]{ppv()}}.
}
\details{
The positive predictive value (\code{\link[=ppv]{ppv()}}) is defined as the percent of
predicted positives that are actually positive while the
negative predictive value (\code{\link[=npv]{npv()}}) is defined as the percent of negative
positives that are actually negative.

Suppose a 2x2 table with notation:

\tabular{rcc}{ \tab Reference \tab \cr Predicted \tab Positive \tab Negative
\cr Positive \tab A \tab B \cr Negative \tab C \tab D \cr }

The formulas used here are:

\deqn{\text{Sensitivity} = \frac{A}{A + C}}

\deqn{\text{Specificity} = \frac{D}{B + D}}

\deqn{\text{Prevalence} = \frac{A + C}{A + B + C + D}}

\deqn{\text{NPV} = \frac{\text{Specificity} \cdot (1 - \text{Prevalence})}{((1 - \text{Sensitivity}) \cdot \text{Prevalence}) + (\text{Specificity} \cdot (1 - \text{Prevalence}))}}

NPV is a metric that should be maximized. The output
ranges from 0 to 1, with
1 indicating all predicted negatives are true
negatives.
}
\section{Relevant Level}{


There is no common convention on which factor level should
automatically be considered the "event" or "positive" result
when computing binary classification metrics. In \code{yardstick}, the default
is to use the \emph{first} level. To alter this, change the argument
\code{event_level} to \code{"second"} to consider the \emph{last} level of the factor the
level of interest. For multiclass extensions involving one-vs-all
comparisons (such as macro averaging), this option is ignored and
the "one" level is always the relevant result.
}

\section{Multiclass}{


Macro, micro, and macro-weighted averaging is available for this metric.
The default is to select macro averaging if a \code{truth} factor with more
than 2 levels is provided. Otherwise, a standard binary calculation is done.
See \code{vignette("multiclass", "yardstick")} for more information.
}

\examples{
# Two class
data("two_class_example")
npv(two_class_example, truth, predicted)

# Multiclass
library(dplyr)
data(hpc_cv)

hpc_cv |>
  filter(Resample == "Fold01") |>
  npv(obs, pred)

# Groups are respected
hpc_cv |>
  group_by(Resample) |>
  npv(obs, pred)

# Weighted macro averaging
hpc_cv |>
  group_by(Resample) |>
  npv(obs, pred, estimator = "macro_weighted")

# Vector version
npv_vec(
  two_class_example$truth,
  two_class_example$predicted
)

# Making Class2 the "relevant" level
npv_vec(
  two_class_example$truth,
  two_class_example$predicted,
  event_level = "second"
)
# Using a different value of 'prevalence'... if you are adding the metric to a
# metric set, you can create a new metric function with the updated argument
# value:

npv_alt_prev  <- metric_tweak("npv_alt_prev", npv, prevalence = 0.40)
multi_metrics <- metric_set(npv, npv_alt_prev)
multi_metrics(two_class_example, truth, estimate = predicted)
}
\references{
Altman, D.G., Bland, J.M. (1994) ``Diagnostic tests 2:
predictive values,'' \emph{British Medical Journal}, vol 309,
102.
}
\seealso{
\link[=class-metrics]{All class metrics}

Other class metrics:
\code{\link[=accuracy]{accuracy()}},
\code{\link[=bal_accuracy]{bal_accuracy()}},
\code{\link[=detection_prevalence]{detection_prevalence()}},
\code{\link[=f_meas]{f_meas()}},
\code{\link[=fall_out]{fall_out()}},
\code{\link[=j_index]{j_index()}},
\code{\link[=kap]{kap()}},
\code{\link[=markedness]{markedness()}},
\code{\link[=mcc]{mcc()}},
\code{\link[=miss_rate]{miss_rate()}},
\code{\link[=ppv]{ppv()}},
\code{\link[=precision]{precision()}},
\code{\link[=recall]{recall()}},
\code{\link[=roc_dist]{roc_dist()}},
\code{\link[=sedi]{sedi()}},
\code{\link[=sens]{sens()}},
\code{\link[=spec]{spec()}}

Other sensitivity metrics:
\code{\link[=fall_out]{fall_out()}},
\code{\link[=miss_rate]{miss_rate()}},
\code{\link[=ppv]{ppv()}},
\code{\link[=sens]{sens()}},
\code{\link[=spec]{spec()}}
}
\author{
Max Kuhn
}
\concept{class metrics}
\concept{sensitivity metrics}
