% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-type-docs.R
\name{numeric-metrics}
\alias{numeric-metrics}
\title{Numeric metrics}
\description{
Numeric metrics evaluate regression predictions where both \code{truth} and
\code{estimate} are numeric. These metrics measure how close predicted values
are to the true values.
}
\section{Input requirements}{

\itemize{
\item \code{truth}: numeric
\item \code{estimate}: numeric
}
}

\section{Available metrics}{

\describe{
\item{\code{\link[=ccc]{ccc()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=gini_coef]{gini_coef()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=huber_loss]{huber_loss()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=huber_loss_pseudo]{huber_loss_pseudo()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=iic]{iic()}}}{Direction: maximize. Range: [-1, 1]}
\item{\code{\link[=mae]{mae()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=mape]{mape()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=mase]{mase()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=mpe]{mpe()}}}{Direction: zero. Range: [-Inf, Inf]}
\item{\code{\link[=msd]{msd()}}}{Direction: zero. Range: [-Inf, Inf]}
\item{\code{\link[=mse]{mse()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=poisson_log_loss]{poisson_log_loss()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=rmse]{rmse()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=rmse_relative]{rmse_relative()}}}{Direction: minimize. Range: [0, Inf]}
\item{\code{\link[=rpd]{rpd()}}}{Direction: maximize. Range: [0, Inf]}
\item{\code{\link[=rpiq]{rpiq()}}}{Direction: maximize. Range: [0, Inf]}
\item{\code{\link[=rsq]{rsq()}}}{Direction: maximize. Range: [-Inf, 1]}
\item{\code{\link[=rsq_trad]{rsq_trad()}}}{Direction: maximize. Range: [0, 1]}
\item{\code{\link[=smape]{smape()}}}{Direction: minimize. Range: [0, 100]}
}
}

\examples{
data("solubility_test")

head(solubility_test)

rmse(solubility_test, solubility, prediction)

}
\seealso{
\link{quantile-metrics} for quantile prediction metrics

\code{vignette("metric-types")} for an overview of all metric types
}
\keyword{internal}
