% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology}
\alias{pathology}
\title{Liver Pathology Data}
\source{
Altman, D.G., Bland, J.M. (1994) ``Diagnostic tests 1:
sensitivity and specificity,'' \emph{British Medical Journal},
vol 308, 1552.
}
\value{
\item{pathology}{a data frame}
}
\description{
Liver Pathology Data
}
\details{
These data have the results of a \emph{x}-ray examination
to determine whether liver is abnormal or not (in the \code{scan}
column) versus the more extensive pathology results that
approximate the truth (in \code{pathology}).
}
\examples{
data(pathology)
str(pathology)
}
\keyword{datasets}
