% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-poisson_deviance.R
\name{poisson_deviance}
\alias{poisson_deviance}
\alias{poisson_deviance.data.frame}
\alias{poisson_deviance_vec}
\title{Mean deviance residuals for Poisson data}
\usage{
poisson_deviance(data, ...)

\method{poisson_deviance}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

poisson_deviance_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate}
columns.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true counts
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:bang-bang]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{poisson_deviance_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the deviance for the Poisson distribution comparing to a saturaed
model.
}
\details{
In general, the deviance statistic is a method for comparing two models
that are fit using maximum likelihood. While the likelihood statistic is
absolute, deviance is defined as a comparative statistic between two models.
In this function, the deviance for a model fit is compared to a model with
perfect fit (i.e., the saturated model).

The value returned is the average of the \emph{deviance residuals} computed using
the function provided by \code{\link[stats:family]{stats::poisson()}}. For zero counts, the residual is
the predicted value. For non-zero counts, the deviance residual is

\deqn{residual = (count * log(count/prediction) - (count - prediction))}

The average of these values is returned.
}
\examples{
count_truth <- c(2L,   7L,   1L,   1L,   0L,  3L)
count_pred  <- c(2.14, 5.35, 1.65, 1.56, 1.3, 2.71)
count_results <- dplyr::tibble(count = count_truth, pred = count_pred)

# Supply truth and predictions as bare column names
poisson_deviance(count_results, count, pred)

}
\seealso{
Other numeric metrics: 
\code{\link{ccc}()},
\code{\link{huber_loss_pseudo}()},
\code{\link{huber_loss}()},
\code{\link{iic}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mase}()},
\code{\link{mpe}()},
\code{\link{msd}()},
\code{\link{poisson_log_loss}()},
\code{\link{rmse}()},
\code{\link{rpd}()},
\code{\link{rpiq}()},
\code{\link{rsq_trad}()},
\code{\link{rsq}()},
\code{\link{smape}()}

Other accuracy metrics: 
\code{\link{ccc}()},
\code{\link{huber_loss_pseudo}()},
\code{\link{huber_loss}()},
\code{\link{iic}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mase}()},
\code{\link{mpe}()},
\code{\link{msd}()},
\code{\link{poisson_log_loss}()},
\code{\link{rmse}()},
\code{\link{smape}()}
}
\author{
Max Kuhn
}
\concept{accuracy metrics}
\concept{numeric metrics}
