% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-poisson_log_loss.R
\name{poisson_log_loss}
\alias{poisson_log_loss}
\alias{poisson_log_loss.data.frame}
\alias{poisson_log_loss_vec}
\title{Mean log loss for Poisson data}
\usage{
poisson_log_loss(data, ...)

\method{poisson_log_loss}{data.frame}(data, truth, estimate, na_rm = TRUE, case_weights = NULL, ...)

poisson_log_loss_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true counts (that is \code{integer}).
This should be an unquoted column name although this argument is passed by
expression and supports \link[rlang:quasiquotation]{quasiquotation} (you can
unquote column names). For \verb{_vec()} functions, an \code{integer} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights. This
should be an unquoted column name that evaluates to a numeric column in
\code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{poisson_log_loss_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the loss function for the Poisson distribution.
}
\details{
Poisson log loss is a metric that should be
minimized. The output ranges from
0 to
Inf, with
0 indicating perfect predictions.

The formula for Poisson log loss is:

\deqn{L = \frac{1}{n} \sum_{i=1}^{n} \left( \log(\text{truth}_i!) + \text{estimate}_i - \text{truth}_i \cdot \log(\text{estimate}_i) \right)}
}
\examples{
count_truth <- c(2L,   7L,   1L,   1L,   0L,  3L)
count_pred  <- c(2.14, 5.35, 1.65, 1.56, 1.3, 2.71)
count_results <- dplyr::tibble(count = count_truth, pred = count_pred)

# Supply truth and predictions as bare column names
poisson_log_loss(count_results, count, pred)

}
\seealso{
\link[=numeric-metrics]{All numeric metrics}

Other numeric metrics:
\code{\link[=ccc]{ccc()}},
\code{\link[=gini_coef]{gini_coef()}},
\code{\link[=huber_loss]{huber_loss()}},
\code{\link[=huber_loss_pseudo]{huber_loss_pseudo()}},
\code{\link[=iic]{iic()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=mase]{mase()}},
\code{\link[=mpe]{mpe()}},
\code{\link[=msd]{msd()}},
\code{\link[=mse]{mse()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmse_relative]{rmse_relative()}},
\code{\link[=rpd]{rpd()}},
\code{\link[=rpiq]{rpiq()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=rsq_trad]{rsq_trad()}},
\code{\link[=smape]{smape()}}

Other accuracy metrics:
\code{\link[=ccc]{ccc()}},
\code{\link[=huber_loss]{huber_loss()}},
\code{\link[=huber_loss_pseudo]{huber_loss_pseudo()}},
\code{\link[=iic]{iic()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=mase]{mase()}},
\code{\link[=mpe]{mpe()}},
\code{\link[=msd]{msd()}},
\code{\link[=mse]{mse()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmse_relative]{rmse_relative()}},
\code{\link[=smape]{smape()}}
}
\author{
Max Kuhn
}
\concept{accuracy metrics}
\concept{numeric metrics}
