% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-type-docs.R
\name{quantile-metrics}
\alias{quantile-metrics}
\title{Quantile metrics}
\description{
Quantile metrics evaluate predictions that consist of predicted quantiles
rather than point predictions. These metrics assess the accuracy and
calibration of distributional forecasts.
}
\section{Input requirements}{

\itemize{
\item \code{truth}: numeric
\item \code{estimate}: \link[hardhat:quantile_pred]{hardhat::quantile_pred} object
}
}

\section{Available metrics}{

\describe{
\item{\code{\link[=weighted_interval_score]{weighted_interval_score()}}}{Direction: minimize. Range: [0, Inf]}
}
}

\examples{
\dontshow{if (rlang::is_installed("hardhat")) withAutoprint(\{ # examplesIf}
library(hardhat)

df <- data.frame(
  preds = quantile_pred(rbind(1:4, 8:11), c(0.2, 0.4, 0.6, 0.8)),
  truth = c(3.3, 7.1)
)

df

weighted_interval_score(df, truth, preds)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{numeric-metrics} for point prediction metrics

\code{vignette("metric-types")} for an overview of all metric types
}
\keyword{internal}
