% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedprob-ranked_prob_score.R
\name{ranked_prob_score}
\alias{ranked_prob_score}
\alias{ranked_prob_score.data.frame}
\alias{ranked_prob_score_vec}
\title{Ranked probability scores for ordinal classification models}
\usage{
ranked_prob_score(data, ...)

\method{ranked_prob_score}{data.frame}(data, truth, ..., na_rm = TRUE, case_weights = NULL)

ranked_prob_score_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected,
and it should correspond to the value of \code{event_level}. Otherwise, there
should be as many columns as factor levels of \code{truth} and the ordering of
the columns should be the same as the factor levels of \code{truth}.}

\item{truth}{The column identifier for the true class results
(that is an \emph{ordered} \code{factor}). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column names).
For \verb{_vec()} functions, a factor vector with class \code{ordered}.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}

\item{estimate}{A matrix with as many columns as factor levels of \code{truth}. \emph{It
is assumed that these are in the same order as the levels of \code{truth}.}}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{ranked_prob_score_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Compute the ranked probability score (RPS) for a classification model using
ordered classes.
}
\details{
Ranked probability score is a metric that should be
minimized. The output ranges from
0 to
1, with 0
indicating perfect predictions.

The ranked probability score is a Brier score for ordinal data that uses the
\emph{cumulative} probability of an event (i.e. \code{Pr[class <= i]} for \code{i} = 1,
2, ..., \code{C} classes). These probabilities are compared to indicators for the
truth being less than or equal to class \code{i}.

Since the cumulative sum of a vector of probability predictions add up to
one, there is an embedded redundancy in the data. For this reason, the raw
mean is divided by the number of classes minus one.

Smaller values of the score are associated with better model performance.
}
\section{Multiclass}{

Ranked probability scores can be computed in the same way for any number of
classes. Because of this, no averaging types are supported.
}

\examples{
library(dplyr)
data(hpc_cv)

hpc_cv$obs <- as.ordered(hpc_cv$obs)

# You can use the col1:colN tidyselect syntax
hpc_cv |>
  filter(Resample == "Fold01") |>
  ranked_prob_score(obs, VF:L)

# Groups are respected
hpc_cv |>
  group_by(Resample) |>
  ranked_prob_score(obs, VF:L)

}
\references{
Wilks, D. S. (2011). \emph{Statistical Methods in the Atmospheric Sciences}.
Academic press. (see Chapter 7)

Janitza, S., Tutz, G., & Boulesteix, A. L. (2016). Random forest for ordinal
responses: prediction and variable selection. Computational Statistics and
Data Analysis, 96, 57-73. (see Section 2)

Lechner, M., & Okasa, G. (2019). Random forest estimation of the ordered
choice model. arXiv preprint arXiv:1907.02436. (see Section 5)
}
\seealso{
\link[=ordered-prob-metrics]{All ordered probability metrics}

Other class probability metrics:
\code{\link[=average_precision]{average_precision()}},
\code{\link[=brier_class]{brier_class()}},
\code{\link[=classification_cost]{classification_cost()}},
\code{\link[=gain_capture]{gain_capture()}},
\code{\link[=mn_log_loss]{mn_log_loss()}},
\code{\link[=pr_auc]{pr_auc()}},
\code{\link[=roc_auc]{roc_auc()}},
\code{\link[=roc_aunp]{roc_aunp()}},
\code{\link[=roc_aunu]{roc_aunu()}}
}
\author{
Max Kuhn
}
\concept{class probability metrics}
