% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-roc_survival_auc.R
\name{roc_survival_auc}
\alias{roc_survival_auc}
\alias{roc_survival_auc.data.frame}
\alias{roc_survival_auc_vec}
\title{ROC Survival AUC}
\usage{
roc_survival_auc(data, ...)

\method{roc_survival_auc}{data.frame}(
  data,
  truth,
  estimate,
  censoring_weights,
  eval_time,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

roc_survival_auc_vec(truth, estimate, censor_probs, .time)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{If \code{truth} is binary, a numeric vector of class probabilities
corresponding to the "relevant" class. Otherwise, a matrix with as many
columns as factor levels of \code{truth}. \emph{It is assumed that these are in the
same order as the levels of \code{truth}.}}

\item{censoring_weights}{The column identifier for censoring weights (that should
be numeric). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{eval_time}{The column identifier for the evaluation time points. This
should be a numeric vector. This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator}, and \code{.estimate}.

For an ungrouped data frame, the result has one row of values. For a grouped data frame,
the number of rows returned is the same as the number of groups.

For \code{roc_survival_auc_vec()}, a \code{numeric} vector same length as the input argument
\code{eval_time}. (or \code{NA}).
}
\description{
Compute the area under the ROC survival curve.
}
\details{
This method automatically groups by the \code{eval_time} argument.

Smaller values of the score are associated with better model performance.
}
\examples{
library(dplyr)

lung_surv \%>\%
  roc_survival_auc(
    truth = surv_obj,
    estimate = .pred_survival,
    censoring_weights = ipcw,
    eval_time = .time
  )
}
\seealso{
Compute the ROC survival curve with \code{\link[=roc_survival_curve]{roc_survival_curve()}}.

Other dynamic survival metrics: 
\code{\link{brier_survival}()}
}
\author{
Emil Hvitfeldt
}
\concept{dynamic survival metrics}
