% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-roc_survival_curve.R
\name{roc_survival_curve}
\alias{roc_survival_curve}
\alias{roc_survival_curve.data.frame}
\title{ROC Survival Curve}
\usage{
roc_survival_curve(data, ...)

\method{roc_survival_curve}{data.frame}(
  data,
  truth,
  estimate,
  censoring_weights,
  eval_time,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{estimate}.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true survival result (that
is created using \code{\link[survival:Surv]{survival::Surv()}}.). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, an \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{estimate}{The column identifier for the survival probabilities (that should be
numeric). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{censoring_weights}{The column identifier for censoring weights (that should
be numeric). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{eval_time}{The column identifier for the evaluation time points. This
should be a numeric vector. This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A tibble with class \code{roc_survival_df}, \code{grouped_roc_survival_df} having
columns \code{.threshold}, \code{recall}, and \code{precision}.
}
\description{
ROC Survival Curve
}
\examples{
roc_survival_curve(
  lung_surv,
  truth = surv_obj,
  estimate = .pred_survival,
  censoring_weights = ipcw,
  eval_time = .time
)
}
\seealso{
Compute the area under the ROC survival curve with (TODO link to
roc_survival_auc()).
}
\author{
Emil Hvitfeldt
}
\concept{survival curve metrics}
