% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-rsq_trad.R
\name{rsq_trad}
\alias{rsq_trad}
\alias{rsq_trad.data.frame}
\alias{rsq_trad_vec}
\title{R squared - traditional}
\usage{
rsq_trad(data, ...)

\method{rsq_trad}{data.frame}(data, truth, estimate, na_rm = TRUE, case_weights = NULL, ...)

rsq_trad_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights. This
should be an unquoted column name that evaluates to a numeric column in
\code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{rsq_trad_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the coefficient of determination using the traditional definition
of R squared using sum of squares. For a measure of R squared that is
strictly between (0, 1), see \code{\link[=rsq]{rsq()}}.
}
\details{
The two estimates for the
coefficient of determination, \code{\link[=rsq]{rsq()}} and \code{\link[=rsq_trad]{rsq_trad()}}, differ by
their formula. The former guarantees a value on (0, 1) while the
latter can generate inaccurate values when the model is
non-informative (see the examples). Both are measures of
consistency/correlation and not of accuracy.

Traditional R squared is a metric that should be
maximized. The output ranges from
0 to 1, with
1 indicating perfect predictions. Negative values
can occur when the model is non-informative.

The formula for traditional R squared is:

\deqn{\text{rsq\_trad} = 1 - \frac{SS_{res}}{SS_{tot}} = 1 - \frac{\sum_{i=1}^{n}(\text{truth}_i - \text{estimate}_i)^2}{\sum_{i=1}^{n}(\text{truth}_i - \bar{\text{truth}})^2}}
}
\examples{
# Supply truth and predictions as bare column names
rsq_trad(solubility_test, solubility, prediction)

library(dplyr)

set.seed(1234)
size <- 100
times <- 10

# create 10 resamples
solubility_resampled <- bind_rows(
  replicate(
    n = times,
    expr = sample_n(solubility_test, size, replace = TRUE),
    simplify = FALSE
  ),
  .id = "resample"
)

# Compute the metric by group
metric_results <- solubility_resampled |>
  group_by(resample) |>
  rsq_trad(solubility, prediction)

metric_results

# Resampled mean estimate
metric_results |>
  summarise(avg_estimate = mean(.estimate))
# With uninformitive data, the traditional version of R^2 can return
# negative values.
set.seed(2291)
solubility_test$randomized <- sample(solubility_test$prediction)
rsq(solubility_test, solubility, randomized)
rsq_trad(solubility_test, solubility, randomized)

}
\references{
Kvalseth. Cautionary note about \eqn{R^2}.
American Statistician (1985) vol. 39 (4) pp. 279-285.
}
\seealso{
\link[=numeric-metrics]{All numeric metrics}

Other numeric metrics:
\code{\link[=ccc]{ccc()}},
\code{\link[=gini_coef]{gini_coef()}},
\code{\link[=huber_loss]{huber_loss()}},
\code{\link[=huber_loss_pseudo]{huber_loss_pseudo()}},
\code{\link[=iic]{iic()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=mase]{mase()}},
\code{\link[=mpe]{mpe()}},
\code{\link[=msd]{msd()}},
\code{\link[=mse]{mse()}},
\code{\link[=poisson_log_loss]{poisson_log_loss()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmse_relative]{rmse_relative()}},
\code{\link[=rpd]{rpd()}},
\code{\link[=rpiq]{rpiq()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=smape]{smape()}}

Other consistency metrics:
\code{\link[=ccc]{ccc()}},
\code{\link[=rpd]{rpd()}},
\code{\link[=rpiq]{rpiq()}},
\code{\link[=rsq]{rsq()}}
}
\author{
Max Kuhn
}
\concept{consistency metrics}
\concept{numeric metrics}
