% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-surv_roc_curve_naive.R
\name{surv_roc_curve_naive}
\alias{surv_roc_curve_naive}
\alias{surv_roc_curve_naive.data.frame}
\alias{surv_roc_curve_naive_vec}
\title{Naive survival receiver operator curve}
\usage{
surv_roc_curve_naive(data, ...)

\method{surv_roc_curve_naive}{data.frame}(
  data,
  truth,
  estimate,
  na_rm = TRUE,
  event_level = yardstick_event_level(),
  ...
)

surv_roc_curve_naive_vec(
  truth,
  estimate,
  na_rm = TRUE,
  event_level = yardstick_event_level(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate} columns.}

\item{...}{Not currently used.}

\item{truth}{A \code{Surv} object. TODO}

\item{estimate}{A list of tibbles. TODO}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{event_level}{A single string. Either \code{"first"} or \code{"second"} to specify
which level of \code{truth} to consider as the "event". This argument is only
applicable when \code{estimator = "binary"}. The default uses an
internal helper that generally defaults to \code{"first"}, however, if the
deprecated global option \code{yardstick.event_first} is set, that will be
used instead with a warning.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
