% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdbrier.R
\name{tdbrier}
\alias{tdbrier}
\alias{tdbrier.model.list}
\alias{tdbrier.int.matrix}
\alias{stdbrier.int.reference}
\alias{get_tdbrier}
\alias{integrate.tdbrier}
\alias{tdbrier.int.reference}
\title{Calculate tdBrier}
\usage{
tdbrier(data, mod, ...)

get_tdbrier(data, mod, ...)

integrate.tdbrier(x, ...)

\method{tdbrier}{int.reference}(x, ...)
}
\arguments{
\item{data}{For the default functions, a datframe containing survival
(time), and status (0:censored/1:event), and the explanatory variables.}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
A tdBrier object
}
\description{
These functions calculate the survival analysis metric measured of a
system compared to a hold-out test set. The measurement and the "truth"
have a survival time and a censoring indicator 0/1 indicating if the event
result or the event.
}
\details{
The Brier score is defined as the squared distance between the
expected survival probability and the observed survival.
Therefore, it measures the discrepancy between observation
and model-based prediction.

The integrated Brier Score summarises the Brier Score over the range
of observed events.Similar to the original Brier score \link{40} the iBrier:
ranges from 0 to 1; the model with an out-of-training sample value closer
to 0 outperforms the rest.
}
\examples{
data("surv_example")

# Given a sample
require(survival)
mod <- coxph(Surv(time, status)~ age, data = surv_example)

tdbrier(surv_example, mod)

}
\references{
Ulla B. Mogensen, Hemant Ishwaran, Thomas A. Gerds (2012).
Evaluating Random Forests for Survival Analysis Using Prediction Error
Curves. Journal of Statistical Software, 50(11), 1-23.
URL http://www.jstatsoft.org/v50/i11/.
}
\seealso{
\link{iBrier}
}
\author{
Carlos S Traynor
}
\keyword{brier}
