% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_example}
\alias{two_class_example}
\title{Two Class Predictions}
\value{
\item{two_class_example}{a data frame}
}
\description{
Two Class Predictions
}
\details{
These data are a test set form a model built for two
classes ("Class1" and "Class2"). There are columns for the true
and predicted classes and column for the probabilities for each
class.
}
\examples{
data(two_class_example)
str(two_class_example)

# truth variable is a two-level factor with the first level set to be
# "Class1" and thus truth == "Class1" is the event of interest for binary
# classification.
levels(hpc_cv$obs)
}
\keyword{datasets}
