% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate-helpers}
\alias{validate-helpers}
\alias{validate_numeric_truth_numeric_estimate}
\alias{validate_factor_truth_factor_estimate}
\alias{validate_factor_truth_matrix_estimate}
\alias{validate_binary_estimator}
\alias{validate_estimator}
\alias{validate_case_weights}
\title{Developer function for validating input in new metrics}
\usage{
validate_numeric_truth_numeric_estimate(truth, estimate)

validate_factor_truth_factor_estimate(truth, estimate)

validate_factor_truth_matrix_estimate(truth, estimate, estimator)

validate_binary_estimator(truth, estimator)

validate_estimator(estimator, estimator_override = NULL)

validate_case_weights(case_weights, size)
}
\arguments{
\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{If \code{truth} is binary, a numeric vector of class probabilities
corresponding to the "relevant" class. Otherwise, a matrix with as many
columns as factor levels of \code{truth}. \emph{It is assumed that these are in the
same order as the levels of \code{truth}.}}

\item{estimator}{One of \code{"binary"}, \code{"hand_till"}, \code{"macro"}, or
\code{"macro_weighted"} to specify the type of averaging to be done. \code{"binary"}
is only relevant for the two class case. The others are general methods for
calculating multiclass metrics. The default will automatically choose
\code{"binary"} if \code{truth} is binary, \code{"hand_till"} if \code{truth} has >2 levels and
\code{case_weights} isn't specified, or \code{"macro"} if \code{truth} has >2 levels and
\code{case_weights} is specified (in which case \code{"hand_till"} isn't
well-defined).}

\item{estimator_override}{A character vector overriding the default allowed
estimator list of \code{c("binary", "macro", "micro", "macro_weighted")}. Set
this if your classification estimator does not support all of these methods.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}

\item{size, }{expected size.}
}
\description{
Developer function for validating input in new metrics
}
\section{Estimator Validation}{

\code{validate_estimator()} is called from your metric specific method of
\code{finalize_estimator_internal()} and ensures that a user provided estimator
is of the right format and is one of the allowed values.
}

