% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-weighted_interval_score.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\alias{weighted_interval_score.data.frame}
\alias{weighted_interval_score_vec}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(data, ...)

\method{weighted_interval_score}{data.frame}(
  data,
  truth,
  estimate,
  quantile_levels = NULL,
  na_rm = TRUE,
  quantile_estimate_nas = c("impute", "drop", "propagate"),
  case_weights = NULL,
  ...
)

weighted_interval_score_vec(
  truth,
  estimate,
  quantile_levels = NULL,
  na_rm = FALSE,
  quantile_estimate_nas = c("impute", "drop", "propagate"),
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the \code{truth}
and \code{estimate} arguments.}

\item{...}{Not Currently used.}

\item{truth}{The column identifier for the true class results
(that is a \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted class results
(that is also \code{quantile_pred}). As with \code{truth} this can be specified
different ways but the primary method is to use an unquoted variable name.
For \verb{_vec()} functions, a \code{quantile_pred} vector.}

\item{quantile_levels}{probabilities. If specified, the score will be
computed at this set of levels. Otherwise, those present in \code{x} will be
used. If \code{quantile_levels} do not exactly match those available in \code{x},
then some quantiles will have implicit missingness. Handling of these is
determined by \code{quantile_estimate_nas}.}

\item{na_rm}{logical. If \code{TRUE}, missing values in \code{actual} or both implicit
and explicit (values of \code{NA} present in \code{x}), will be ignored (dropped) in
the calculation of the summary score. If \code{FALSE} (the default), any \code{NA}s
will result in the summary being \code{NA}.}

\item{quantile_estimate_nas}{character. This argument applies only to \code{x}.
It handles imputation of individual \code{quantile_levels} that are necessary to
compute a score. Because each element of \code{x} is a \link[hardhat:quantile_pred]{hardhat::quantile_pred},
it is possible for these to be missing for particular
\code{quantile_levels}. There are a number of different possibilities for such
missingness. The options are as follows:
\itemize{
\item For \code{"impute"}, both explicit and implicit missing values will be imputed
using \code{\link[hardhat:impute_quantiles]{hardhat::impute_quantiles()}} prior to the calculation of the score.
So the score will be \code{NA} only if imputation fails.
\item For \code{"drop"}, any explicit missing values will be removed
before calculating the score for a particular prediction. This may be
reasonable due to the weighting. For example, if the estimate has
\code{quantile_levels = c(.25, .5, .75)} but the median is \code{NA} for a particular
prediction, it may be reasonable to average the accuracy of \code{c(.25, .75)}
for that prediction with others that don't have missingness. This option
is only works if \code{quantile_levels = NULL} or is a subset of the
\code{quantile_levels} in \code{x}.
\item For \code{"propagate"}, any missing value predictions will result in that
element of \code{x} having a score of \code{NA}. If \code{na_rm = TRUE}, then these will
be removed before averaging.
}}

\item{case_weights}{The optional column identifier for case weights. This
should be an unquoted column name that evaluates to a numeric column in
\code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}
}
\value{
a vector of nonnegative scores.
}
\description{
Weighted interval score (WIS), a well-known quantile-based approximation of
the commonly-used continuous ranked probability score (CRPS). WIS is a proper
score, and can be thought of as a distributional generalization of absolute
error. For example, see
\href{https://arxiv.org/abs/2005.12881}{Bracher et al. (2020)} for discussion in
the context of COVID-19 forecasting.
}
\details{
Weighted interval score is a metric that should be
minimized. The output ranges from
0 to
Inf, with
0 indicating perfect predictions.
}
\examples{
library(hardhat)

quantile_levels <- c(.2, .4, .6, .8)
pred1 <- 1:4
pred2 <- 8:11
preds <- quantile_pred(rbind(pred1, pred2), quantile_levels)
truth <- c(3.3, 7.1)
weighted_interval_score_vec(truth, preds)
weighted_interval_score_vec(truth, preds, quantile_levels = c(.25, .5, .75))

# Missing value behaviours

preds_na <- quantile_pred(rbind(pred1, c(1, 2, NA, 4)), 1:4 / 5)
truth <- c(2.5, 2.5)
weighted_interval_score_vec(truth, preds_na)
weighted_interval_score_vec(truth, preds_na, quantile_levels = 1:9 / 10)
try(weighted_interval_score_vec(
  truth,
  preds_na,
  quantile_levels = 1:9 / 10,
  quantile_estimate_nas = "drop"
))
weighted_interval_score_vec(
  truth,
  preds_na,
  quantile_levels = c(2, 3) / 5,
  quantile_estimate_nas = "drop"
)
weighted_interval_score_vec(
  truth, preds_na, na_rm = TRUE, quantile_estimate_nas = "propagate"
)
weighted_interval_score_vec(
  truth, preds_na, quantile_estimate_nas = "propagate"
)

}
\seealso{
\link[=quantile-metrics]{All quantile metrics}
}
