% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-package.R
\docType{package}
\name{yardstick-package}
\alias{yardstick}
\alias{yardstick-package}
\title{yardstick: Tidy Characterizations of Model Performance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tidy tools for quantifying how well model fits to a data set such as confusion matrices, class probability curve summaries, and regression metrics (e.g., RMSE).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/yardstick}
  \item \url{https://yardstick.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/yardstick/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})
  \item Max Kuhn \email{max@posit.co}
  \item Davis Vaughan \email{davis@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
