% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missings.R
\name{yardstick_remove_missing}
\alias{yardstick_remove_missing}
\alias{yardstick_any_missing}
\title{Developer function for handling missing values in new metrics}
\usage{
yardstick_remove_missing(truth, estimate, case_weights)

yardstick_any_missing(truth, estimate, case_weights)
}
\arguments{
\item{truth}{The realized vector of \code{truth}. This is either a factor or a
numeric.}

\item{estimate}{The realized \code{estimate} result. This is either a numeric
vector, a factor vector, or a numeric matrix (in the case of multiple class
probability columns) depending on your metric function.}

\item{case_weights}{The realized case weights, as a numeric vector. This must
be the same length as \code{truth}, and will be considered in the \code{na_rm}
checks. This will be passed on to \code{metric_impl} as the named argument
\code{case_weights}.}
}
\description{
\code{yardstick_remove_missing()},  and \code{yardstick_any_missing()} are useful alongside the
\link{metric-summarizers} functions for implementing new custom metrics.
\code{yardstick_remove_missing()} removes any observations that contains missing values
across, truth, estimate and case_weights. \code{yardstick_any_missing()} returns an
indicator if there is any missing values in the inputs.
}
\seealso{
\link{metric-summarizers}
}
