% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{accuracy}
\alias{accuracy}
\alias{accuracy.data.frame}
\alias{accuracy.table}
\alias{accuracy.matrix}
\alias{kap}
\alias{kap.data.frame}
\alias{kap.table}
\alias{kap.matrix}
\title{Classification Metrics on Predited Classes}
\usage{
accuracy(data, ...)

\method{accuracy}{data.frame}(data, truth, estimate, na.rm = TRUE, ...)

\method{accuracy}{table}(data, ...)

\method{accuracy}{matrix}(data, ...)

kap(data, ...)

\method{kap}{data.frame}(data, truth, estimate, na.rm = TRUE, ...)

\method{kap}{table}(data, ...)

\method{kap}{matrix}(data, ...)
}
\arguments{
\item{data}{For the default functions, a factor containing the
discrete measurements. For the \code{table} or \code{matrix}
functions, a table or matrix object, respectively, where the
true class results should be in the columns of the table.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a factor). This should an unquoted column name although
this argument is passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names or column positions).}

\item{estimate}{The column identifier for the predicted class
results (that is also factor). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be stripped before the computation proceeds}
}
\value{
A numeric value.
}
\description{
Accuracy is the proportion of the data that are
predicted correctly. Kappa is a similar measure but is normalized by
the accuracy that would be expected by chance alone and is very useful
when one or more classes have large frequency distributions.
}
\examples{
library(dplyr)
data("hpc_cv")

# The observed and predicted classes from a single
# assessment set (i.e. fold)
fold_1 <- hpc_cv \%>\%
  filter(Resample == "Fold01")

fold_1 \%>\% conf_mat(truth = obs, estimate = pred)

fold_1 \%>\% accuracy(truth = obs, estimate = pred)

fold_1 \%>\% kap(truth = obs, estimate = pred)

fold_1 \%>\% metrics(truth = obs, estimate = pred)
}
\references{
Cohen, J. (1960). "A coefficient of agreement for nominal
scales". \emph{Educational and Psychological Measurement}. 20 (1): 37–46.
}
\seealso{
\code{\link[=conf_mat]{conf_mat()}}, \code{\link[=metrics]{metrics()}}
}
\author{
Max Kuhn
}
\keyword{manip}
