% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-brier.R
\name{brier_survival}
\alias{brier_survival}
\alias{brier_survival.data.frame}
\alias{brier_survival_vec}
\title{Brier score for right censored data}
\usage{
brier_survival(data, ...)

\method{brier_survival}{data.frame}(
  data,
  truth,
  estimate,
  censoring_weights,
  .time,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

brier_survival_vec(
  truth,
  estimate,
  censoring_weights,
  .time,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{estimate}.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class survival result (that
is created using \code{\link[survival:Surv]{survival::Surv()}}.). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, an \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{estimate}{The column identifier for the survival probabilities. This is
expected as a list of tibbles, containing 2 columns \code{.time} and
\code{.pred_survival}. This should be an unquoted column name although this
argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a list of tibbles, each containing 2 columns \code{.time} and
\code{.pred_survival}.}

\item{censoring_weights}{The column identifier for censoring weights. This is
expected to a numeric vector. This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{.time}{A vector of time points.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values. \code{.estimate} will be a list column of
tibbles. Each containing 2 columns \code{.time} and \code{.estimate}, with the number
of rows corresponding to the input argument \code{.time}.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{brier_survival_vec()}, a \code{numeric} vector same length as the input argument
\code{.time}. (or \code{NA}).
}
\description{
Compute the time-dependent Brier score for right censored data. Which is the
mean squared error at time point \code{.time}.
}
\details{
Smaller values of the score are associated with better model performance.
}
\examples{
res <- brier_survival(
  data = lung_surv,
  truth = surv_obj,
  estimate = .pred,
  censoring_weights = prob_censored,
  .time = c(100, 500, 1000)
)

res

res[[".estimate"]]
}
\references{
E. Graf, C. Schmoor, W. Sauerbrei, and M. Schumacher, “Assessment and
comparison of prognostic classification schemes for survival data,”
Statistics in Medicine, vol. 18, no. 17-18, pp. 2529–2545, 1999.
}
\author{
Emil Hvitfeldt
}
\concept{dynamic survival metrics}
