% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num-ccc.R
\name{ccc}
\alias{ccc}
\alias{ccc.data.frame}
\alias{ccc_vec}
\title{Concordance correlation coefficient}
\usage{
ccc(data, ...)

\method{ccc}{data.frame}(data, truth, estimate, bias = FALSE,
  na_rm = TRUE, ...)

ccc_vec(truth, estimate, bias = FALSE, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate}
columns.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \code{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \code{_vec()} functions, a \code{numeric} vector.}

\item{bias}{A \code{logical}; should the biased estimate of variance
be used (as is Lin (1989))?}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{ccc_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Calculate the concordance correlation coefficient.
}
\details{
\code{\link[=ccc]{ccc()}} is a metric of both consistency/correlation and accuracy,
while metrics such as \code{\link[=rmse]{rmse()}} are strictly for accuracy and metrics
such as \code{\link[=rsq]{rsq()}} are strictly for consistency/correlation
}
\examples{
# Supply truth and predictions as bare column names
ccc(solubility_test, solubility, prediction)

library(dplyr)

set.seed(1234)
size <- 100
times <- 10

# create 10 resamples
solubility_resampled <- bind_rows(
  replicate(
    n = times,
    expr = sample_n(solubility_test, size, replace = TRUE),
    simplify = FALSE
  ),
  .id = "resample"
)

# Compute the metric by group
metric_results <- solubility_resampled \%>\%
  group_by(resample) \%>\%
  ccc(solubility, prediction)

metric_results

# Resampled mean estimate
metric_results \%>\%
  summarise(avg_estimate = mean(.estimate))
}
\references{
Lin, L. (1989). A concordance correlation
coefficient to evaluate reproducibility. \emph{Biometrics}, 45 (1),
255–268.

Nickerson, C. (1997). A note on "A concordance correlation
coefficient to evaluate reproducibility". \emph{Biometrics}, 53(4),
1503-1507.
}
\seealso{
Other numeric metrics: \code{\link{mae}},
  \code{\link{mape}}, \code{\link{rmse}},
  \code{\link{rpd}}, \code{\link{rpiq}},
  \code{\link{rsq_trad}}, \code{\link{rsq}},
  \code{\link{smape}}

Other consistency metrics: \code{\link{rpd}},
  \code{\link{rpiq}}, \code{\link{rsq_trad}},
  \code{\link{rsq}}

Other accuracy metrics: \code{\link{mae}},
  \code{\link{mape}}, \code{\link{rmse}},
  \code{\link{smape}}
}
\author{
Max Kuhn
}
\concept{accuracy metrics}
\concept{consistency metrics}
\concept{numeric metrics}
