% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-metric.R
\name{check_metric}
\alias{check_metric}
\alias{check_numeric_metric}
\alias{check_class_metric}
\alias{check_prob_metric}
\title{Developer function for checking inputs in new metrics}
\usage{
check_numeric_metric(truth, estimate, case_weights)

check_class_metric(truth, estimate, case_weights, estimator)

check_prob_metric(truth, estimate, case_weights, estimator)
}
\arguments{
\item{truth}{The realized vector of \code{truth}. This is either a factor or a
numeric.}

\item{estimate}{The realized \code{estimate} result. This is either a numeric
vector, a factor vector, or a numeric matrix (in the case of multiple class
probability columns) depending on your metric function.}

\item{case_weights}{The realized case weights, as a numeric vector. This must
be the same length as \code{truth}, and will be considered in the \code{na_rm}
checks. This will be passed on to \code{metric_impl} as the named argument
\code{case_weights}.}

\item{estimator}{This can either be \code{NULL} for the default auto-selection of
averaging (\code{"binary"} or \code{"macro"}), or a single character to pass along to
the metric implementation describing the kind of averaging to use.}
}
\description{
\code{check_numeric_metric()}, \code{check_class_metric()}, and \code{check_prob_metric()}
are useful alongside the \code{\link[=metric-summarizers]{metric-summarizers()}} functions for implementing
new custom metrics. \code{\link[=metric-summarizers]{metric-summarizers()}} calls the metric function inside
\code{dplyr::summarise()}. These functions perform checks for the input in
accordance to the type of metric that is used.
}
\seealso{
\code{\link[=metric-summarizers]{metric-summarizers()}}
}
