% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-concordance.R
\name{concordance_survival}
\alias{concordance_survival}
\alias{concordance_survival.data.frame}
\alias{concordance_survival_vec}
\title{Concordance index for right-censored data}
\usage{
concordance_survival(data, ...)

\method{concordance_survival}{data.frame}(
  data,
  truth,
  estimate,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

concordance_survival_vec(
  truth,
  estimate,
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{estimate}.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class survival result (that
is created using \code{\link[survival:Surv]{survival::Surv()}}.). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, an \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{estimate}{The column identifier for the survival probabilities. This
needs to be a numeric vector.This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{concordance_survival_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Compute the Concordance index for right-censored data
}
\details{
The concordance index is defined as the proportion of all comparable pairs in
which the predictions and outcomes are concordant.
}
\examples{
concordance_survival(data = lung_surv, truth = surv_obj, estimate = age)
}
\author{
Emil Hvitfeldt
}
\concept{static survival metrics}
