% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_mat.R
\name{conf_mat}
\alias{conf_mat}
\alias{conf_mat.table}
\alias{conf_mat.default}
\alias{conf_mat.data.frame}
\alias{conf_mat.table}
\title{Confusion Matrix for Categorical Data}
\usage{
conf_mat(data, ...)

\method{conf_mat}{data.frame}(data, truth = NULL, estimate = NULL,
  dnn = c("Prediction", "Truth"), ...)

\method{conf_mat}{table}(data, ...)
}
\arguments{
\item{data}{A data frame or a \code{\link[base:table]{base::table()}}.}

\item{...}{Options to pass to \code{\link[base:table]{base::table()}} (not including \code{dnn})}

\item{truth}{A string corresponding to the column in the data frame containig  the factor of classes to be used as the true results}

\item{estimate}{A string corresponding to the column in the data frame containig  the factor of predicted classes.}

\item{dnn}{a character vector of dimnames for the table}
}
\description{
Calculates a cross-tabulation of observed and predicted classes with
associated statistics.
}
\details{
The functions requires that the factors have exactly the same levels.
}
