% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mcc.R
\name{mcc}
\alias{mcc}
\alias{mcc.data.frame}
\alias{mcc_vec}
\title{Matthews correlation coefficient}
\usage{
mcc(data, ...)

\method{mcc}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

mcc_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{Either a \code{data.frame} containing the \code{truth} and \code{estimate}
columns, or a \code{table}/\code{matrix} where the true class results should be
in the columns of the table.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \code{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \code{_vec()} functions, a \code{factor} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{mcc_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Matthews correlation coefficient
}
\section{Relevant level}{


There is no common convention on which factor level should
automatically be considered the "event" or "positive" results.
In \code{yardstick}, the default is to use the \emph{first} level. To
change this, a global option called \code{yardstick.event_first} is
set to \code{TRUE} when the package is loaded. This can be changed
to \code{FALSE} if the last level of the factor is considered the
level of interest.
}

\section{Multiclass}{


\code{mcc()} has a known multiclass generalization and that is computed
automatically if a factor with more than 2 levels is provided. Because
of this, no averaging methods are provided.
}

\examples{
# Two class
data("two_class_example")
mcc(two_class_example, truth, predicted)

# Multiclass
# mcc() has a natural multiclass extension
library(dplyr)
data(hpc_cv)
hpc_cv \%>\%
  group_by(Resample) \%>\%
  mcc(obs, pred)

}
\references{
Giuseppe, J. (2012). "A Comparison of MCC and CEN Error
Measures in Multi-Class Prediction". \emph{PLOS ONE}. Vol 7, Iss 8, e41882.
}
\seealso{
Other class metrics: \code{\link{accuracy}},
  \code{\link{bal_accuracy}},
  \code{\link{detection_prevalence}}, \code{\link{f_meas}},
  \code{\link{j_index}}, \code{\link{kap}},
  \code{\link{npv}}, \code{\link{ppv}},
  \code{\link{precision}}, \code{\link{recall}},
  \code{\link{sens}}, \code{\link{spec}}
}
\author{
Max Kuhn
}
\concept{class metrics}
