% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_set}
\alias{metric_set}
\title{Combine metric functions}
\usage{
metric_set(...)
}
\arguments{
\item{...}{The bare names of the functions to be included in the metric set.}
}
\description{
\code{metric_set()} allows you to combine multiple metric functions together
into a new function that calculates all of them at once.
}
\details{
There are currently no checks in place to ensure that all of the metric
functions calculate the same kind of metric (regression vs classification).
}
\examples{

library(dplyr)

# Multiple regression metrics
multi_metric <- metric_set(rmse, rsq, ccc)

multi_metric(solubility_test, truth = solubility, estimate = prediction)

# Groups are respected on the new metric function
class_metrics <- metric_set(accuracy, kap)

hpc_cv \%>\%
  group_by(Resample) \%>\%
  class_metrics(obs, pred)

# If you need to set options for certain metrics,
# do so ahead of time with purrr::partial()
library(purrr)
ccc_with_bias <- partial(ccc, bias = TRUE)

multi_metric2 <- metric_set(rmse, rsq, ccc_with_bias)

multi_metric2(solubility_test, truth = solubility, estimate = prediction)

}
\seealso{
\code{\link[=metrics]{metrics()}}
}
