% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{metrics.data.frame}
\title{General Function to Estimate Performance}
\usage{
metrics(data, ...)

\method{metrics}{data.frame}(data, truth, estimate, ...,
  options = list(), na.rm = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate}
columns and any columns specified by \code{...}.}

\item{...}{For \code{mnLogLoss()}, a set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. There should be as many columns as
factor levels of \code{truth}. It is \strong{assumed} that they are in the
same order as the factor levels. Otherwise, unused.}

\item{truth}{The column identifier for the true results (that
is \code{numeric} or \code{factor}). This should be an unquoted column name
although this argument is passed by expression and support
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names).}

\item{estimate}{The column identifier for the predicted results
(that is also \code{numeric} or \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name.}

\item{options}{A \code{list} of named options to pass to \code{\link[=roc]{roc()}}
such as \code{direction} or \code{smooth}. These options should not include \code{response},
\code{predictor}, or \code{levels}.}

\item{na.rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
A two column tibble.
\itemize{
\item When \code{truth} is a factor, there are columns for \code{\link[=accuracy]{accuracy()}} and the
Kappa statistic (\code{\link[=kap]{kap()}}).
\item If a full set of class probability columns are passed to \code{...}, then
there is also a column for \code{\link[=mnLogLoss]{mnLogLoss()}}.
\item When \code{truth} has two levels and there are class probabilities, \code{\link[=roc_auc]{roc_auc()}}
is appended.
\item When \code{truth} is numeric, there are columns for \code{\link[=rmse]{rmse()}}, \code{\link[=rsq]{rsq()}},
and \code{\link[=mae]{mae()}}.
}
}
\description{
This function estimates one or more common performance
estimates depending on the class of \code{truth} (see \strong{Value}
below) and returns them in a two column tibble.
}
\examples{

# Accuracy and kappa
metrics(two_class_example, truth, predicted)

# Add on multinomal log loss and ROC AUC by specifying class prob columns
metrics(two_class_example, truth, predicted, Class1, Class2)

# Regression metrics
metrics(solubility_test, truth = solubility, estimate = prediction)

}
\seealso{
\code{\link[=metric_set]{metric_set()}}
}
