% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{roc_auc}
\alias{roc_auc}
\alias{roc_auc.default}
\alias{pr_auc}
\alias{pr_auc.default}
\alias{roc_auc.data.frame}
\alias{pr_auc}
\alias{pr_auc.data.frame}
\alias{mnLogLoss}
\alias{mnLogLoss.data.frame}
\title{Metrics Based on Class Probabilities}
\usage{
roc_auc(data, ...)

\method{roc_auc}{data.frame}(data, truth = NULL, estimate = NULL,
  na.rm = TRUE, ...)

pr_auc(data, ...)

\method{pr_auc}{data.frame}(data, truth = NULL, estimate = NULL,
  na.rm = TRUE, ...)

mnLogLoss(data, ...)

\method{mnLogLoss}{data.frame}(data, truth = NULL, estimate = NULL,
  na.rm = TRUE, sum = FALSE, ...)
}
\arguments{
\item{data}{A data frame with the relevant columns.}

\item{...}{Options to pass to \code{\link[=roc]{roc()}} such as \code{direction} or
\code{smooth}. These options should not include \code{response},
\code{predictor}, or \code{levels}. No options are available to pass to
\code{\link[=PRAUC]{PRAUC()}}.}

\item{truth}{A single character value containing the column
name of \code{data} that contains the true classes (in a factor).}

\item{estimate}{A character vector containing the columns
name of \code{data} that contain the predicted class probabilities
or some other score (in numeric vectors). These values are
assumed to have larger values associated with the event,
although this can be changed by passing the \code{direction} argument
to \code{\link[=roc]{roc()}} via the \code{...} when computing the ROC curve.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be stripped before the computation proceeds}

\item{sum}{A logical. Should the sum of the likelihood
contrinbutions be returned (instead of the mean value)?}
}
\value{
A number between 0 and 1 (or NA) for \code{roc_auc} or
\code{pr_auc}. For \code{mnLogLoss} a number of \code{NA}.
}
\description{
These functions compute the areas under the receiver operating
characteristic (ROC) curve (\code{roc_auc}), the precision-recall
curve (\code{pr_auc}), or the multinomial log loss (\code{mnLogLoss}).
}
\details{
There is no common convention on which factor level should
automatically be considered the "relevant" or "positive" results.
In \code{yardstick}, the default is to use the \emph{first} level. To
change this, a global option called \code{yardstick.event_first} is
set to \code{TRUE} when the package is loaded. This can be changed
to \code{FALSE} if the last level of the factor is considered the
level of interest.
}
\examples{
data("two_class_example")
prob_cols <- levels(two_class_example$truth)

roc_auc(two_class_example, truth = "truth", estimate = prob_cols)
roc_auc(two_class_example, truth = "truth", estimate = prob_cols,
        smooth = TRUE)
        
pr_auc(two_class_example, truth = "truth", estimate = prob_cols)    

mnLogLoss(two_class_example, truth = "truth", estimate = prob_cols)            
}
\seealso{
\code{\link[=conf_mat]{conf_mat()}}, \code{\link[=summary.conf_mat]{summary.conf_mat()}}, \code{\link[=recall]{recall()}}, \code{\link[=mcc]{mcc()}}
}
\keyword{manip}
