% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-roc_auc_survival.R
\name{roc_auc_survival}
\alias{roc_auc_survival}
\alias{roc_auc_survival.data.frame}
\alias{roc_auc_survival_vec}
\title{ROC Survival AUC}
\usage{
roc_auc_survival(data, ...)

\method{roc_auc_survival}{data.frame}(data, truth, ..., na_rm = TRUE, case_weights = NULL)

roc_auc_survival_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true survival result (that
is created using \code{\link[survival:Surv]{survival::Surv()}}.). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, an \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}

\item{estimate}{If \code{truth} is binary, a numeric vector of class probabilities
corresponding to the "relevant" class. Otherwise, a matrix with as many
columns as factor levels of \code{truth}. \emph{It is assumed that these are in the
same order as the levels of \code{truth}.}}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator}, and \code{.estimate}.

For an ungrouped data frame, the result has one row of values. For a grouped data frame,
the number of rows returned is the same as the number of groups.

For \code{roc_auc_survival_vec()}, a \code{numeric} vector same length as the input argument
\code{eval_time}. (or \code{NA}).
}
\description{
Compute the area under the ROC survival curve.
}
\details{
Smaller values of the score are associated with better model performance.
}
\examples{
library(dplyr)

lung_surv \%>\%
  roc_auc_survival(
    truth = surv_obj,
    .pred
  )
}
\seealso{
Compute the ROC survival curve with \code{\link[=roc_curve_survival]{roc_curve_survival()}}.

Other dynamic survival metrics: 
\code{\link{brier_survival_integrated}()},
\code{\link{brier_survival}()}
}
\author{
Emil Hvitfeldt
}
\concept{dynamic survival metrics}
