% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-roc_curve_survival.R
\name{roc_curve_survival}
\alias{roc_curve_survival}
\alias{roc_curve_survival.data.frame}
\title{ROC Survival Curve}
\usage{
roc_curve_survival(data, ...)

\method{roc_curve_survival}{data.frame}(data, truth, ..., na_rm = TRUE, case_weights = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{The column identifier for the survival probabilities this
should be a list column of data.frames corresponding to the output given when
predicting with {censored} model. This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}

\item{truth}{The column identifier for the true survival result (that
is created using \code{\link[survival:Surv]{survival::Surv()}}.). This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, an \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector.}
}
\value{
A tibble with class \code{roc_survival_df}, \code{grouped_roc_survival_df} having
columns \code{.threshold}, \code{recall}, and \code{precision}.
}
\description{
ROC Survival Curve
}
\examples{
result <- roc_curve_survival(
  lung_surv,
  truth = surv_obj,
  .pred
)
result

#' # ---------------------------------------------------------------------------
# `autoplot()`

# Visualize the curve using ggplot2 manually
library(ggplot2)
library(dplyr)
result \%>\%
  ggplot(aes(x = 1 - specificity, y = sensitivity)) +
  geom_path() +
  geom_abline(lty = 3) +
  coord_equal() +
  theme_bw()

# Or use autoplot
autoplot(result)
}
\seealso{
Compute the area under the ROC survival curve with \code{\link[=roc_auc_survival]{roc_auc_survival()}}.
}
\author{
Emil Hvitfeldt
}
\concept{survival curve metrics}
