% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-royston.R
\name{royston_survival}
\alias{royston_survival}
\title{Royston-Sauerbei D statistic}
\usage{
royston_survival(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{Currently not used.}

\item{estimate}{The column identifier for the predicted linear predictor,
this should be a numeric variable. This should be an unquoted column name
although this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column names). For
\verb{_vec()} functions, a numeric vector.}
}
\value{
A \code{tibble} with columns \code{.metric}, \code{.estimator},
and \code{.estimate} and 1 row of values.

For grouped data frames, the number of rows returned will be the same as
the number of groups.

For \code{royston_survival_vec()}, a single \code{numeric} value (or \code{NA}).
}
\description{
Compute the Royston-Sauerbei D statistic
}
\details{
Royston and Sauerbrei proposed $R^2_D$ as a measure of explained variation
on the log relative hazard scale based on the authors’ D statistic.
D measures prognostic separation of survival curves, and is closely related
to the standard deviation of the prognostic index.

Larger values of the score are associated with better model performance.
}
\examples{
royston_survival(
  data = lung_surv,
  truth = surv_obj,
  estimate = .pred_linear_pred
)
}
\references{
Royston, P., Sauerbrei, W., "A new measure of prognostic separation in
survival data", Statistics in Medicine, 23, 723-748, 2004.
}
\author{
Hannah Frick
}
\concept{linear pred survival metrics}
