% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-extra.R
\name{rpd}
\alias{rpd}
\alias{rpd.data.frame}
\alias{rpd_vec}
\alias{rpiq}
\alias{rpiq.data.frame}
\alias{rpiq_vec}
\title{Extra Metrics for Numeric Outcomes}
\usage{
rpd(data, ...)

\method{rpd}{data.frame}(data, truth, estimate, na.rm = TRUE, ...)

rpd_vec(truth, estimate, na.rm = TRUE, ...)

rpiq(data, ...)

\method{rpiq}{data.frame}(data, truth, estimate, na.rm = TRUE, ...)

rpiq_vec(truth, estimate, na.rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the \code{truth} and \code{estimate}
columns.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names). For \code{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \code{_vec()} functions, a \code{numeric} vector.}

\item{na.rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\value{
For \code{_vec()} functions, a single \code{numeric} value (or \code{NA}).
Otherwise, a \code{tibble} with columns \code{.metric} and \code{.estimate} and 1 row of
values. For grouped data frames, the number of rows returned will be the
same as the number of groups.
}
\description{
These functions are appropriate for cases where the model
outcome is a number. The ratio of performance to deviation
(\code{rpd()}) and the ratio of performance to inter-quartile (\code{rpiq()})
are both measures of consistency/correlation between observed
and predicted values (and not of accuracy).
}
\details{
In the field of spectroscopy in particular, the ratio
of performance to deviation (RPD) has been used as the standard
way to report the quality of a model. It is the ratio between
the standard deviation of a variable and the standard error of
prediction of that variable by a given model. However, its
systematic use has been criticized by several authors, since
using the standard deviation to represent the spread of a
variable can be misleading on skewed dataset. The ratio of
performance to inter-quartile has been introduced by
Bellon-Maurel et al. (2010) to address some of these issues, and
generalise the RPD to non-normally distributed variables.
}
\examples{

rpd(solubility_test, truth = solubility, estimate = prediction)
rpiq(solubility_test, truth = solubility, estimate = prediction)

}
\references{
Williams, P.C. (1987) Variables affecting near-infrared
reflectance spectroscopic analysis. In: Near Infrared Technology
in the Agriculture and Food Industries. 1st Ed. P.Williams and
K.Norris, Eds. Am. Cereal Assoc. Cereal Chem., St. Paul, MN.

Bellon-Maurel, V., Fernandez-Ahumada, E., Palagos, B., Roger,
J.M. and McBratney, A., (2010). Critical review of chemometric
indicators commonly used for assessing the quality of the
prediction of soil attributes by NIR spectroscopy. TrAC Trends
in Analytical Chemistry, 29(9), pp.1073-1081.
}
\author{
Pierre Roudier
}
\keyword{manip}
