% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sens}
\alias{sens}
\alias{sens.default}
\alias{sens.table}
\alias{sens.matrix}
\alias{spec}
\alias{spec.default}
\alias{spec.table}
\alias{spec.matrix}
\alias{ppv}
\alias{ppv.default}
\alias{ppv.table}
\alias{ppv.matrix}
\alias{npv}
\alias{npv.default}
\alias{npv.table}
\alias{npv.matrix}
\alias{sens.data.frame}
\alias{sens.table}
\alias{sens.matrix}
\alias{spec.data.frame}
\alias{ppv}
\alias{ppv.table}
\alias{ppv.matrix}
\alias{npv}
\alias{npv.table}
\alias{npv.matrix}
\title{Calculate sensitivity, specificity and predictive values}
\usage{
sens(data, ...)

\method{sens}{data.frame}(data, truth = NULL, estimate = NULL,
  na.rm = TRUE, ...)

\method{sens}{table}(data, ...)

\method{sens}{matrix}(data, ...)

\method{spec}{data.frame}(data, truth = NULL, estimate = NULL,
  na.rm = TRUE, ...)

ppv(data, ...)

\method{ppv}{table}(data, prevalence = NULL, ...)

\method{ppv}{matrix}(data, prevalence = NULL, ...)

npv(data, ...)

\method{npv}{table}(data, prevalence = NULL, ...)

\method{npv}{matrix}(data, prevalence = NULL, ...)
}
\arguments{
\item{data}{For the default functions, a factor containing the
discrete measurements. For the \code{table} or \code{matrix}
functions, a table or matrix object, respectively.}

\item{...}{Not currently used.}

\item{truth}{A single character value containing the column
name of \code{data} that contains the true classes (in a factor).}

\item{estimate}{A single character value containing the column
name of \code{data} that contains the predicted classes (in a factor).}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be stripped before the computation proceeds}

\item{prevalence}{A numeric value for the rate of the
"positive" class of the data.}
}
\value{
A number between 0 and 1 (or NA).
}
\description{
These functions calculate the sensitivity, specificity or
predictive values of a measurement system compared to a
reference results (the truth or a gold standard). The
measurement and "truth" data must have the same two possible
outcomes and one of the outcomes must be thought of as a
"positive" results or the "event".
}
\details{
The sensitivity is defined as the proportion of positive
results out of the number of samples which were actually
positive. When there are no positive results, sensitivity is not
defined and a value of \code{NA} is returned. Similarly, when
there are no negative results, specificity is not defined and a
value of \code{NA} is returned. Similar statements are true for
predictive values.

The positive predictive value is defined as the percent of
predicted positives that are actually positive while the
negative predictive value is defined as the percent of negative
positives that are actually negative.

TODO mention prob interpretations

Suppose a 2x2 table with notation

\tabular{rcc}{ \tab Reference \tab \cr Predicted \tab Event \tab No Event
\cr Event \tab A \tab B \cr No Event \tab C \tab D \cr }

The formulas used here are: \deqn{Sensitivity = A/(A+C)} \deqn{Specificity =
D/(B+D)} \deqn{Prevalence = (A+C)/(A+B+C+D)} \deqn{PPV = (sensitivity *
Prevalence)/((sensitivity*Prevalence) + ((1-specificity)*(1-Prevalence)))}
\deqn{NPV = (specificity * (1-Prevalence))/(((1-sensitivity)*Prevalence) +
((specificity)*(1-Prevalence)))}

See the references for discussions of the statistics.
}
\examples{


}
\references{
Altman, D.G., Bland, J.M. (1994) ``Diagnostic tests 1:
sensitivity and specificity,'' \emph{British Medical Journal},
vol 308, 1552.

Altman, D.G., Bland, J.M. (1994) ``Diagnostic tests 2:
predictive values,'' \emph{British Medical Journal}, vol 309,
102.
}
\seealso{
\code{\link[=conf_mat]{conf_mat()}}
}
\author{
Max Kuhn
}
\keyword{manip}
