% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdroc.R
\name{tdroc}
\alias{tdroc}
\alias{get_tdroc}
\alias{integrate.tdroc}
\title{Calculate tdROC}
\usage{
tdroc(data, mod, ...)

get_tdroc(data, mod, ...)

integrate.tdroc(mod, ...)
}
\arguments{
\item{data}{For the default functions, a datframe containing survival
(time), and status (0:censored/1:event), and the explanatory variables.}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
A tdROC object
}
\description{
The time dependent ROC analysis (tdROC) allows
the user to visualise the sensitivity (true positive rate) and
specificity (false positive rate) for all possible cut-offs for the
predicted survival.
#'
}
\examples{
data("surv_example")

# Given a sample
require(survival)
mod <- coxph(Surv(time, status)~ age, data = surv_example)

tdroc(surv_example, mod)

}
\references{
Liang Li, Cai Wu Department of Biostatistics and
The University of Texas MD Anderson Cancer Center (2016).
tdROC: Nonparametric Estimation of Time-Dependent ROC Curve
from Right Censored Survival Data. R package version 1.0.
https://CRAN.R-project.org/package=tdROC
}
\seealso{
\link{iROC}
}
\author{
Carlos S Traynor
}
\keyword{tdroc}
