% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_example}
\alias{two_class_example}
\title{Two Class Predictions}
\value{
\item{two_class_example}{a data frame}
}
\description{
Two Class Predictions
}
\details{
These data are a test set form a model built for two
classes ("Class1" and "Class2"). There are columns for the true
and predicted classes and column for the probabilities for each
class.
}
\examples{
data(two_class_example)
str(two_class_example)

# `truth` is a 2 level factor. The first level is `"Class1"`, which is the
# "event of interest" by default in yardstick. See the Relevant Level
# section in any classification function (such as `?pr_auc`) to see how
# to change this.
levels(hpc_cv$obs)
}
\keyword{datasets}
