% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-package.R
\docType{package}
\name{yardstick-package}
\alias{yardstick}
\alias{yardstick-package}
\title{yardstick: Tidy Characterizations of Model Performance}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tidy tools for quantifying how well model fits to a data set such as confusion matrices, class probability curve summaries, and regression metrics (e.g., RMSE).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/yardstick}
  \item \url{https://yardstick.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/yardstick/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
