% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{yardstick_remove_missing}
\alias{yardstick_remove_missing}
\alias{yardstick_any_missing}
\title{Developer function for handling missing values in new metrics}
\usage{
yardstick_remove_missing(truth, estimate, case_weights)

yardstick_any_missing(truth, estimate, case_weights)
}
\arguments{
\item{truth, estimate}{Vectors of the same length.}

\item{case_weights}{A vector of the same length as \code{truth} and \code{estimate}, or
\code{NULL} if case weights are not being used.}
}
\description{
\code{yardstick_remove_missing()},  and \code{yardstick_any_missing()} are useful
alongside the \link{metric-summarizers} functions for implementing new custom
metrics. \code{yardstick_remove_missing()} removes any observations that contains
missing values across, \code{truth}, \code{estimate} and \code{case_weights}.
\code{yardstick_any_missing()} returns \code{FALSE} if there is any missing values in
the inputs.
}
\seealso{
\link{metric-summarizers}
}
